fix + gọi api tin tức, quảng cáo
