+++ src/services/index.ts
@@ -6,3 +6,4 @@ export * from './payment.service';
 export * from './notification.service';
 export * from './staff.service';
 export * from './clinicService.service';
+export * from './news.service';
