+++ src/services/news.service.ts
@@ -1,0 +1,8 @@
+import { IApiResponse, INews, INewsQueryParams, IPagination } from '@/types';
+import { axiosClient } from '@/utils';
+
+export const newsApi = {
+    getNews(params: INewsQueryParams): Promise<IApiResponse<IPagination<INews>>> {
+        return axiosClient.get(`/news`, { params: params });
+      },
+};
