+++ src/services/staff.service.ts
@@ -1,9 +1,13 @@
-import { IApiResponse } from '@/types';
+import { IApiResponse, IClinicStaffDetail } from '@/types';
 import { IClinicStaff, IStaffQueryParams } from '@/types';
 import { axiosClient } from '@/utils';
 
 export const staffApi = {
-  getStaffs(params: IStaffQueryParams): Promise<IApiResponse<IClinicStaff[]>> {
-    return axiosClient.get('/staffs', { params: params });
+  getClinicStaffs(params: IStaffQueryParams): Promise<IApiResponse<IClinicStaff[]>> {
+    return axiosClient.get(`/staffs`, { params: params });
+  },
+
+  getClinicStaff(staffId: string): Promise<IApiResponse<IClinicStaffDetail>> {
+    return axiosClient.get(`/staffs/${staffId}`);
   },
 };
