+++ src/types/clinic.types.ts
@@ -17,15 +17,6 @@ export interface IClinicQueryParams {
   isActive?: boolean;
 }
 
-export interface IStaffQueryParams{
-  userId?: string;
-  roleId?: string;
-  clinicId?: string;
-  gender?: number;
-  phoneNumber?: string;
-  email?: string;
-  name?: string;
-}
 
 export interface IAddClinicService {
   id: number;
@@ -135,40 +126,4 @@ export interface IInviteClinicMemberRequest {
   email: string;
   roleId: number;
   clinicId: string;
-}
-
-export interface IClinicStaff{
-  id: number;
-  experience?: number;
-  description?: string;
-  specialize?: string;
-  users: IClinicStaffDetail;
-  role: IUserGroupRole;
-}
-
-export interface IClinicStaffDetail {
-  id?: number;
-  memberId?: number;
-  specialize?: string;
-  experience?: number;
-  phoneNumber?: string;
-  address?: string;
-  gender?: number;
-  createdAt?: string;
-  updatedAt?: string;
-  isDisabled?: boolean;
-  avatar?: string;
-  disabledAt?: string;
-  userId?: string;
-  clinicId?: string;
-  name?: string;
-  firstName?: string;
-  lastName?: string;
-}
-
-export interface IClinicService{
-  id: number;
-  serviceName?: string;
-  price?: number;
-  description?: string;
 }
