+++ src/types/news.type.ts
@@ -1,0 +1,18 @@
+export interface INewsQueryParams{
+    clinicId?: string;
+    title?: string;
+    isShow?: boolean;
+    pageSize?: string;
+    pageIndex?: string;
+  }
+
+export interface INews {
+    id: number;
+    title?: string;
+    content?: string;
+    logo?: string;
+    clinicId: string;
+    isShow: boolean;
+    createdAt?: string;
+    updatedAt?: string;
+  }
